refactor: Extract and organize files into domain-oriented structure
