+++ backend/src/services/mod.rs
@@ -1,1 +1,2 @@
-pub mod part_service;
+// pub mod part_service;
+pub mod part;
