+++ backend/src/services/part/mod.rs
@@ -1,0 +1,9 @@
+pub mod create;
+pub mod delete;
+pub mod get;
+pub mod update;
+
+pub use create::create_part;
+pub use delete::delete_part;
+pub use get::{get_part, get_parts};
+pub use update::update_part;
